backend/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Koichi Hirachi <kouichi.hirachi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
openai = "^0.28.1"
pyaudio = "^0.2.13"
tiktoken = "^0.5.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
data-url = "^1.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
